AuthAPI/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
